refactored and added a readme

+++ src/main/kotlin/com/kooken/kotlincourse/lesson15/Wind.kt
@@ -1,8 +1,8 @@
 package com.kooken.kotlincourse.lesson15
 
-class Wind (val speed: Int){
+class Wind(val speed: Int) {
 
-    fun convertSpeed (): Int{
-        return speed * 1000 / 3600
+    fun convertSpeed(): Double {
+        return speed * 1000.0 / 3600
     }
 }
